FIX: remove argo barrier and use node-wide acquire
To make sure that all threads have finished sampling, we make use
of the implicit synchronization imposed by the MPI_Allreduce calls.
For this reason, we just need to invoke a node-wide acquire to see
the changes performed by the rest of the processes.

TODO: We don't need to invalidate everything, but only what we need.