Use Argo's locality info to sample
This commit introduces the CMake option `-DUSE_LOCALITY`(set ON/OFF).
When this is set to `ON`, under the ArgoDSM version, each node under
`sample` will only compute the items that has local in its memory ba-
nk. With that, we completely eliminate write misses! This is achieved
using Argo's `get_homenode` API call. That said, the first-touch allo-
cation policy can/should not be used as the data is not touched during
initialization.

NOTE: Under `-DUSE_LOCALITY=ON`, `-DARGO_RELEASE` can (better) be left
to `NO_RELEASE` as the release operations will result to no ops.

`USE_LOCALITY` is by default set to `ON`.