Adjuct self-invalidation to Argo's locality implementation
If we choose to compute based on ArgoDSM's locality information,
we render useless the connectivity matrix. For this reason, we
have to invalidate based on ArgoDSM's locality as well.

TODO: The most optimal approach would be to create the connecti-
vity matrix taking into account ArgoDSM's locality. This way we
will both preserve locality and invalidate only what we need to.