Change coherence ops order, add verification option, and more
This commit changes the order of how the coherence operations are
done. It also adds an MPI node-wide barrier after the releaes ops
as the MPI_Allreduce calls, previously used as an implicit barrier,
are not guaranteed to stall all processes.

Furthermore, this commit adds a CMake verification option when we
want to check for correctness, as when benchmarking we would like
to avoid the additional coherence operations that come along with
the verification. The option is `ENABLE_VERIFICATION` (default ON).

Some compiler checks are also added just in case the coherence
operations combination of choice violates correctness.