Move `acquire` to the place it initially was
Since this is an iterative benchmark, the overhead we might incur
having the computational acquire() plus the verification acquire()
is significant. As in this commit, the computational acquire() se-
rves also the verification acquire.

The order still remains:
-- acquire
-- compute
-- release

as the computational part is preceded by a node-wide Argo barrier.

We also avoid specifying `in.` in all of the acquire coherence fu-
nctions, which dirty the code.

If this change is still unwanted, revert to:
bfd13bf288e9a23434e4476e3b8cfc350aee7554