Remove the locality information implementation
This commit completely remotes the `ENABLE_LOCALITY` option and tou-
ches the Argo data correctly before going into the actual execution
through the `argo_data_init` function. It is essential that the data
is initialized through that function and not in execution, as remote
data not residing in a node's chunk might be touched otherwise.

With this change, the changes in sample() and predict() are reverted.
Also, code related to `ENABLE_LOCALITY` is removed under bpmf_argo.h.